ContactForm customization:
I was trying to avoid monkeypatching by extending the ContactForm model,
view, and controller and then routing to the extended classes. But the
form was always hitting the controller empty for some reason and it
became too time-intensive to make it work that way.

So I'm deleting the extended classes and rerouting and just patching the
hyrax classes like usual.